Save system overhaul, introduced Kayako compiler
Completely reworked save system to avoid having to reference lines of code. Previously, the lines were saved which would break the save file if any new lines were added to the script. The save system now relies entirely on page names and labels. It only creates a save state when you call doLink() or if you call doChoice() when the choice has an associated label. The history works independently of the save state, so you can rewind to previous choices without labels, but the game won't save there.

Similarly, + choices that are meant to be single use only work if there is a label associated with in.

The story parsing functions were completely removed from sadako.js and instead placed inside kayako.js. Kayako will be the name of the compiler and its functions.

A command-line compiler was also created with kayako_cli.js. It can be run or compiled with node.js.

sadako variable definitions were moved outside of init() so that they can be accessed without the need to call init().

Implimented version checking. The version will be saved to the story_data member of the JSON file on compile and checked against current version on loading. It'll throw an error if it's older.

Added abort() and end() functions that are the javascript equivelant of calling "<< ABORT" and "<< END".