Added Kayako versioning. Modified loading of sadako.story in init()
Made a separate version for Kayako. The version checking is only for determining whether the data in the JSON file would be formatted incorrectly. It shouldn't be ncessary to recompile your script unless Kayako has changed in some way, and so the version check will be against Kayako's vesion number, not Sadako's. Therefore,  having two version numbers is ideal.

Modified the loading of sadako.story in init() to be less error prone.