Bug fixes. (See description)
Modified random functions.

Added  updateData() which is called from loadData() that merges the loaded data with the default data, so as to not remove new data that might not be in an older save file.

Added sadako.in_dialog boolean to determine whether the dialog window is open or not.

Switched from story lines be displayed a <p> tags to <div> tags.

Added setJump() and getJump()  functions which push/pop the jump and the current conditions to an array. Not storing conditions on jump was causing errors.

Increased doLines() loop from 20 to 200 because 20 was not enough.

Adjusted how the after functions in doScript() were handled so that it only calls them if it's at the very end of page rendering.